Removing old backup file + Directory placeholders
This was a backup file that I created during the move to .org files for
Emacs config. There was a point where I thought that I lost a bunch of
uncommitted work (in Emacs), but it still seemed to be in the file. So I
backed up the file so that it wouldn't get overwritten.

Also adding some some directory placeholders (.gitignore files) that
ensure that the directories exist on a new clone of the repo.